+++ src/Control/Effect.hs
@@ -3,6 +3,7 @@ module Control.Effect
 ) where
 
 import Control.Effect.Carrier   as X ((:+:), Carrier, Effect, Member)
+import Control.Effect.Choose    as X (Choose, ChooseC)
 import Control.Effect.Cull      as X (Cull, CullC, OnceC)
 import Control.Effect.Cut       as X (Cut, CutC)
 import Control.Effect.Empty     as X (Empty, EmptyC)